Added more shared interfaces
