+++ src/shared/types.ts
@@ -27,12 +27,33 @@ export interface IProblem {
   buildCommand: string;
 }
 
+export interface IProblemBase {
+  _id?: string;
+  title: string;
+  // add more basic details as necessary
+}
+
 /** Module interfaces */
 
 export interface IModuleBase {
   _id?: string;
   name: string;
   number: number;
+}
+
+/** Feedback Interfaces */
+
+/** Useful Enum */
+export enum AlertType {
+  info = "info",
+  error = "error",
+  success = "success",
+}
+
+export interface IFeedback {
+  message?: string;
+  display: boolean;
+  type: AlertType;
 }
 
 /** Other Interfaces */
